renamed travel-journal-screen.tsx to profile-screen.tsx to better reflect updated changes.
post button on the create post screen also properly redirects to the default profile page now (completed last commit).
